DAO 1.2.2 + Service 1.0

+++ ticket-system/src/main/java/com/tsystems/javaschool/vm/sub/Hasher.java
@@ -36,9 +36,9 @@ public class Hasher {
         return ((long) (Math.random() * Long.MAX_VALUE));
     }
 
-    public static int [] reverseArray(int[] array) {
+    public static long [] reverseArray(long[] array) {
         for (int i = 0; i < array.length / 2; i++) {
-            int tmp = array[array.length - 1 - i];
+            long tmp = array[array.length - 1 - i];
             array[array.length - 1 - i] = array[i];
             array[i] = tmp;
         }
